Initial commit for data persistence.
