+++ flexibility/LinkDef.h
@@ -1,0 +1,15 @@
+#ifdef __ROOTCLING__
+
+#pragma link off all globals;
+#pragma link off all classes;
+#pragma link off all functions;
+
+#pragma link C++ nestedclasses;
+#pragma link C++ nestedtypedef;
+
+#pragma link C++ class BaseCell+;
+#pragma link C++ class Neurite+;
+#pragma link C++ class Neuron<BaseCell>+;
+#pragma link C++ class NeuronExtension<Neuron<BaseCell> >+;
+
+#endif
