+++ flexibility/neurite.cc
@@ -1,0 +1,5 @@
+#include "neurite.h"
+
+Neurite::Neurite() {}
+
+Neurite::~Neurite() {}
